requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
psycopg2-binary==2.9.9
supabase==2.3.4
boto3==1.34.0
requests==2.31.0
